feat: complete file organizer script

+++ organizer.sh
@@ -2,9 +2,9 @@
 
 read -p "Please enter the directory path: " dir_path
 
-if [ ! -d "$dir_path" ] ;then
-	echo "Error.. The "$dir_path" directory does not exist"
-	exit 1
+if [ ! -d "$dir_path" ]; then
+    echo "Error.. The '$dir_path' directory does not exist"
+    exit 1
 fi
 
 echo "Directory found! Proceeding to organize..."
@@ -14,35 +14,59 @@ mkdir -p "$dir_path/Documents"
 mkdir -p "$dir_path/Videos"
 mkdir -p "$dir_path/Archives"
 
-for file in "$dir_path"/*;
-do
-	if [ -f "$file" ]; then
+count_images=0
+count_docs=0
+count_videos=0
+count_archives=0
+count_others=0
 
-		filename=$(basename "$file")
-		if [[ "$filename" == *.jpg ]] || [[ "$filename" == *.jpeg ]] || [[ "$filename" == *.png ]] || [[ "$filename" == *.gif ]] ; then
-			dest_folder="Images"
+date >> "$dir_path/log.txt"
 
-		elif [[ "$filename" == *.pdf ]] || [[ "$filename" == *.doc ]] ||[[ "$filename" == *.docx ]] || [[ "$filename" == *.txt ]] ; then
-			dest_folder="Documents"
+for file in "$dir_path"/*; do
+    if [ -f "$file" ]; then
+        
+        filename=$(basename "$file")
+        
+        if [[ "$filename" != "log.txt" ]]; then
 
-		elif [[ "$filename" == *.mp4 ]] || [[ "$filename" == *.mkv ]] ||[[ "$filename" == *.avi ]] ; then
-			dest_folder="Videos"
+            if [[ "$filename" == *.jpg ]] || [[ "$filename" == *.jpeg ]] || [[ "$filename" == *.png ]] || [[ "$filename" == *.gif ]]; then
+                dest_folder="Images"
+                count_images=$((count_images + 1))
 
-		elif [[ "$filename" == *.zip ]] || [[ "$filename" == *.rar ]] ||[[ "$filename" == *.tar.gz ]] ; then
-			dest_folder="Archives"
+            elif [[ "$filename" == *.pdf ]] || [[ "$filename" == *.doc ]] || [[ "$filename" == *.docx ]] || [[ "$filename" == *.txt ]]; then
+                dest_folder="Documents"
+                count_docs=$((count_docs + 1))
 
-		else 
-			dest_folder="Others"
-			mkdir -p "$dir_path/$dest_folder"
-		fi
+            elif [[ "$filename" == *.mp4 ]] || [[ "$filename" == *.mkv ]] || [[ "$filename" == *.avi ]]; then
+                dest_folder="Videos"
+                count_videos=$((count_videos + 1))
 
-		final_name="$filename"
-       	 	while [ -e "$dir_path/$dest_folder/$final_name" ]; do
-            		echo "Warning: A file named '$final_name' already exists in $dest_folder."
-            		read -p "Please enter a new name for this file (including extension): " final_name
-        	done
+            elif [[ "$filename" == *.zip ]] || [[ "$filename" == *.rar ]] || [[ "$filename" == *.tar.gz ]]; then
+                dest_folder="Archives"
+                count_archives=$((count_archives + 1))
 
-		mv "$file" "$dir_path/$dest_folder/$final_name"
-	        echo "Moved: $filename -> $dest_folder/$final_name"
-	fi
+            else
+                dest_folder="Others"
+                mkdir -p "$dir_path/$dest_folder"
+                count_others=$((count_others + 1))
+            fi
+
+            final_name="$filename"
+            while [ -e "$dir_path/$dest_folder/$final_name" ]; do
+                read -p "File $final_name exists! enter new name: " final_name
+            done
+
+            mv "$file" "$dir_path/$dest_folder/$final_name"
+            
+            echo "Moved: $filename -> $dest_folder/$final_name"
+            echo "$filename -> $dest_folder/$final_name" >> "$dir_path/log.txt"
+            
+        fi
+    fi
 done
+
+echo "Images: $count_images"
+echo "Documents: $count_docs"
+echo "Videos: $count_videos"
+echo "Archives: $count_archives"
+echo "Others: $count_others"
